Move tests to nightly
