+++ tests/nightly/tg/ccl/test_llama_prefill_ccl_ops_perf_nightly_TG.py
@@ -1,0 +1,1 @@
+../../../ttnn/unit_tests/operations/ccl/test_llama_prefill_ccl_ops_perf_nightly_TG.py
